cleaned up codes and added home page tasks script
