+++ public/home.html
@@ -120,8 +120,8 @@
                 <main-banner>
                     <account-banner>
                         <div>Well-List</div>
-                        <div>Username</div>
-                        <div>Logout</div>
+                        <div id="username-label">Username</div>
+                        <div id="logout-button">Logout</div>
                     </account-banner>
                     <top-half id="date-banner"></top-half>
                     <bottom-half>
