+++ public/scripts/home-page/main.js
@@ -1,4 +1,5 @@
 import { initializeMainBanner } from "./main-banner.js";
+import { initializeTaskSection } from "./tasks.js";
 import { loadRewardsSection } from "./rewards-section.js";
 
 const focusedDate = new Date();
