+++ public/scripts/home-page/tasks.js
@@ -1,0 +1,3 @@
+export function initializeTaskSection() {
+    
+}
