+++ public/scripts/rewards-shop/right-menu.js
@@ -1,6 +1,6 @@
 import * as data from '../local-data.js';
 import * as sprites from '../sprite-sources.js';
-import * as utils from './utils.js';
+import * as utils from '../utils.js';
 import * as constants from './right-menu-constants.js';
 import { handleThemeChange } from './main.js';
